backend/pyproject.toml
[project]
name = "arxgorithm-backend"
version = "0.1.0"
description = "Open-source personalized arXiv paper recommendation API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.34.0,<1.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "asyncpg>=0.30.0,<1.0",
    "pgvector>=0.4,<1.0",
    "argon2-cffi>=23.1,<26.0",
    "PyJWT>=2.8,<3.0",
    "python-multipart>=0.0.18,<1.0",
    "email-validator>=2.0,<3.0",
    "numpy>=1.26,<3.0",
    "openai>=1.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10.0",
    "pytest-asyncio>=0.24,<2.0",
    "ruff>=0.8,<1.0",
    "mypy>=1.13,<2.0",
]

[build-system]
requires = ["hatchling>=1.29.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
